backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0.post1
pydantic==2.5.2
pydantic-settings==2.1.0
langchain==0.2.16
langchain-groq==0.1.9
langchain-community==0.2.16
sentence-transformers==2.5.1
pinecone-client==3.0.0
firebase-admin==6.3.0
celery==5.3.6
redis==5.0.1
PyPDF2==3.0.1
beautifulsoup4==4.12.2
requests==2.31.0
python-multipart==0.0.6
python-dotenv==1.0.0
lxml==5.1.0
httpx==0.27.0
